+++ crates/cubek-tile/src/tile/view/mod.rs
@@ -1,4 +1,5 @@
 pub mod accumulate;
+pub mod coords;
 pub mod flat;
 pub mod masked;
 pub mod matrix;
@@ -6,6 +7,8 @@ pub mod projected;
 pub mod quant;
 
 pub use accumulate::*;
+// Crate-internal helpers, so this re-export carries no public item.
+pub(crate) use coords::*;
 pub use flat::*;
 pub use masked::*;
 pub use matrix::*;
